Bet sizing test 2x faster, warning fixes

+++ mlfinlab/bet_sizing/ch10_snippets.py
@@ -85,7 +85,7 @@ def mp_avg_active_signals(signals, molecule):
     :param molecule: (list) Indivisible tasks to be passed to 'mp_pandas_obj', in this case a list of datetimes.
     :return: (pandas.Series) The averaged bet size sub-series.
     """
-    out = pd.Series()
+    out = pd.Series(dtype='float64')
     for loc in molecule:
         df0 = (signals.index.to_numpy() <= loc)&((loc < signals['t1'])|pd.isnull(signals['t1']))
         act = signals[df0].index
